% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqplot.R
\name{lambda}
\alias{lambda}
\title{Compute genomic inflation factor lambda}
\usage{
lambda(p)
}
\arguments{
\item{pvector}{Vector of p-values}
}
\value{
Value of λ
}
\description{
This function computes the genomic inflation factor (λ) using the median approach.
}
