% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan_plot.R
\name{manhattan_plot}
\alias{manhattan_plot}
\title{Manhattan plot}
\usage{
manhattan_plot(
  probe,
  region = NULL,
  array = c("450K", "EPIC"),
  FDR = FALSE,
  title = NULL,
  col.chr = c("gray40", "gray55"),
  col.sig = "gray18",
  col.dmr = "black",
  line.dmr = "#4393c3",
  size.line.dmr = 0.2,
  cex = 0.5,
  cex.sig = 0.75,
  alpha = 0.5,
  alpha.sig = 1,
  size.line.sig = 0.5
)
}
\arguments{
\item{probe}{Dataframe of EWAS results, including the columns 'cpg', 'P.Value', 'chr', and 'pos'. If 'cpg' column is not included, rownames will be used as probe names. If output from limma's topTable is provided, chromosome and position will be obtained from 450K of EPIC annotations.}

\item{region}{Dataframe of regional analysis results including the columns 'chr', 'start', 'end'.}

\item{array}{'450K' or 'EPIC'}

\item{FDR}{Should an FDR level of significance be indicated? Defaults to FALSE.}

\item{title}{Title text.}

\item{col.chr}{Vector of two colors that will be alternately used for probes within each chromosome. Defaults to c('gray40', 'gray55').}

\item{col.sig}{Color of FDR or Bonferroni significant probes. Defaults to 'gray18'.}

\item{col.dmr}{Color of probes within DMRs. Defaults to 'black'.}

\item{line.dmr}{Color of line indicating DMRs. Defaults to '#4393c3'.}

\item{size.line.dmr}{Size of line indicating DMRs. Defaults to 0.2.}

\item{cex}{Size of non-significant probes. Defaults to 0.5.}

\item{cex.sig}{Size of significant probes. Defaults to 0.5.}

\item{alpha}{Alpha of non-significant probes. Defaults to 0.5.}

\item{alpha.sig}{alpha Alpha of non-significant probes. Defaults to 1.}

\item{Size}{of line indicating FDR and Bonferroni significance. Defaults to 0.5.}
}
\value{
Q-Q plot
}
\description{
This function creates a Manhattan plot of EWAS results using ggplot2.
}
