% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq_plot.R
\name{qq_plot}
\alias{qq_plot}
\title{Plot Q-Q plot}
\usage{
qq_plot(pvector)
}
\arguments{
\item{pvector}{Vector of p-values}
}
\value{
Q-Q plot
}
\description{
This function creates a quantile-quantile (Q-Q) plot of p-values using ggplot2.
}
