% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqplot.R
\name{qqplot}
\alias{qqplot}
\title{Plot Q-Q plot}
\usage{
qqplot(pvector)
}
\arguments{
\item{pvector}{Vector of p-values}
}
\value{
Q-Q plot
}
\description{
This function creates a quantile-quantile (Q-Q) plot of p-values using ggplot2.
}
